+++ CHANGELOG.md
@@ -1,12 +1,11 @@
 # Changelog
 
 # Todo
-- wasm - add wasm-functions dbBufferExport(), dbBufferImport()
 - sqlmath - add sql-extension carrayblob()
 - none
 
 # v2022.6.1-beta
-- functional index.html
+- sqlmath - milestone - functional index.html
 - add files asset_sqlmath_external_rollup.css, asset_sqlmath_external_rollup.js
 - bugfix - fix memory-leak by replacing all free()/malloc() with sqlite3_free()/sqlite3_malloc()
 - wasm - update functions dbFileExportAsync(), dbFileImportAsync() to be able to export/import raw-db-arraybuffer